Refactored some of the startup code
I have refactored the startup code, but it is still coupled with other pages and the code as well. I will decouple and put all of the startup code into its own folder. For now, the card nugget and writeup portion has been done. Now is to refactor the higher overview of the code, which is the page and the body itself